LPS-195528 Add the ability to delete an FDS action, and extract the form to its own component to separate state between components
